import chocolatey helpers to make refreshenv available in powershell
This fixes the error:

```
Update-SessionEnvironment : The term 'Update-SessionEnvironment' is not recognized as the name of a cmdlet, function,
script file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is
correct and try again.
At C:\tmp\vagrant-shell.ps1:13 char:1
+ Update-SessionEnvironment
+ ~~~~~~~~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : ObjectNotFound: (Update-SessionEnvironment:String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException
```